Remove hasLocal and use more direct rules
Instead of declaring a fact that a class may store the result of a call in
a variable, add a rule to set the variable directly. This is more efficient
and allows the rule to depend on the value in some way.

Also, creation with "new" is also modelled as a call-site.