New config {} syntax
Instead of using datalog to describe the initial objects and their connections,
a Java-like syntax can now be used.

Internally, this creates a new _testDriver object instance of a _TestDriver class.
First, all the "setup" blocks are run in a copy of the model. Then, all created
objects and their fields are copied to the real model and the "test" blocks are
run.

Some related changes:

- The position of arguments is now taken into account
- Graph nodes can be hidden (used to hide _testDriver)
- New objects are named based on the variable they are stored in, not their type
- A new phase/1 relation can be used to do phase-specific actions
- Integer literals are now allowed
- Added didCall/3 to simplify writing goals